\name{acmacs.PlotSpec}
\title{Class to access and modify plot specification of the chart.}
\alias{acmacs.PlotSpec-class}
\alias{acmacs.PlotSpec}
\docType{package}
\usage{
ps <- chart$plot_spec
}
\value{An object of class \code{"acmacs.PlotSpec"} to access and modify
  plot specification of the chart. It has the following properties and methods:

  \item{name}{name of the chart.}


}
\keyword{ package }
\examples{
}
\seealso{
  \code{\link{acmacs.Chart}}\cr
  \code{\link{acmacs.Antigen}}\cr
  \code{\link{acmacs.Serum}}\cr
  \code{\link{acmacs.Projection}}\cr
  \code{\link{acmacs.Example}}
}
\author{Eugene Skepner, eu@antigenic-cartography.org.}
